Refactor so JsonOperation implements UnaryOperation and separate all the TableViz code in TableVizOperation.
