+++ distribution/lib/Standard/Table/0.0.0-dev/src/Internal/In_Memory_Column_Implementation.enso
@@ -581,8 +581,9 @@ type In_Memory_Column_Implementation
                 apply_unary_map this_column this_column.name formatter Value_Type.Char on_problems=..Report_Error
 
     json_stringify (this_column : Column & In_Memory_Column) =
-        new_column = JsonOperation.INSTANCE.makeJsonColumn this_column.java_column
-        In_Memory_Column.from_java_column new_column
+        Java_Problems.with_problem_aggregator on_problems java_problem_aggregator->
+            new_column = JsonOperation.INSTANCE.apply this_column.java_column java_problem_aggregator
+            In_Memory_Column.from_java_column new_column
 
     cast (this_column : Column & In_Memory_Column) value_type:Value_Type on_problems:Problem_Behavior =
         Cast_Helpers.check_cast_compatibility this_column.value_type value_type
