+++ distribution/lib/Standard/Table/0.0.0-dev/src/Row.enso
@@ -6,6 +6,7 @@ import project.Errors.No_Such_Column
 import project.Internal.Widget_Helpers
 from project.Internal.Storage import java_to_enso
 
+polyglot java import org.enso.table.data.column.operation.TableVizOperation
 polyglot java import org.enso.table.data.table.Row as Java_Row # necessary for reflection metadata
 
 ## Represents a single row of some in-memory Table.
@@ -100,7 +101,7 @@ type Row
        ---
        Converts this row into a JSON representation.
     to_json_data self -> Text =
-        self.java_row.toJsonData
+        TableVizOperation.makeJSONForRow self.java_row
 
 ## ---
    private: true
