+++ std-bits/table/src/main/java/org/enso/table/data/table/Row.java
@@ -2,7 +2,6 @@ package org.enso.table.data.table;
 
 import java.util.Arrays;
 import java.util.function.Function;
-import org.enso.table.data.column.operation.JsonOperation;
 
 public class Row {
   private final Table table;
@@ -33,6 +32,10 @@ public class Row {
     return table.getColumns().length;
   }
 
+  public String get_name(int index) {
+    return table.getColumns()[index].getName();
+  }
+
   public Object get_value(int index, Function<Object, Object> ifMissing) {
     var count = column_count();
     if (index < -count || index >= count) {
@@ -47,22 +50,5 @@ public class Row {
   public Object get_value(String name, Function<Object, Object> ifMissing) {
     var column = table.getColumnByName(name);
     return column == null ? ifMissing.apply(name) : column.getItem(rowIndex);
-  }
-
-  public String toJsonData() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("{");
-    for (int i = 0; i < column_count(); i++) {
-      if (i > 0) {
-        sb.append(",");
-      }
-      String name = table.getColumns()[i].getName();
-      Object value = get_value(i, null);
-      sb.append(JsonOperation.VIZ_INSTANCE.objectToJson(name))
-          .append(":")
-          .append(JsonOperation.VIZ_INSTANCE.objectToJson(value));
-    }
-    sb.append("}");
-    return sb.toString();
   }
 }
